Converted useGetLocalToken and useGetLocalUser hooks to global helpers
