--- src/layouts/auth-layout/hooks/useGetLocalToken.ts
@@ -1,8 +1,0 @@
-import LocalStorageKeys from '../../../constants/localStorageKeys';
-
-const useGetLocalToken = (): string | null => {
-  const token = localStorage.getItem(LocalStorageKeys.TOKEN);
-  return token;
-};
-
-export default useGetLocalToken;
