+++ src/utils/getLocalToken.ts
@@ -1,0 +1,8 @@
+import LocalStorageKeys from '../constants/localStorageKeys';
+
+const getLocalToken = (): string | null => {
+  const token = localStorage.getItem(LocalStorageKeys.TOKEN);
+  return token;
+};
+
+export default getLocalToken;
