+++ src/utils/getLocalUser.ts
@@ -1,0 +1,10 @@
+import { LocalUser } from '../types/user';
+import LocalStorageKeys from '../constants/localStorageKeys';
+
+const getLocalUser = (): LocalUser | null => {
+  const user = localStorage.getItem(LocalStorageKeys.USER);
+  if (!user) return null;
+  return JSON.parse(user);
+};
+
+export default getLocalUser;
